pyproject.toml
[tool.poetry]
name = "animalshelter"
version = "0.1.0"
description = ""
authors = ["Aikanysh <lailievaa02@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
pillow = "^10.1.0"
environs = "^10.0.0"
mysqlclient = "^2.2.1"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
